+++ src/main/java/com/and1/view/label/HistogramLabelGray.java
@@ -1,4 +1,4 @@
-package com.and1.gui.label;
+package com.and1.view.label;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
