+++ src/main/java/com/and1/view/label/HistogramLabelRGB.java
@@ -1,4 +1,4 @@
-package com.and1.gui.label;
+package com.and1.view.label;
 
 import java.awt.*;
