+++ src/main/java/com/and1/view/renderer/ImageCellRendererSorted.java
@@ -1,14 +1,14 @@
-package com.and1.gui.renderer;
+package com.and1.view.renderer;
 
 import com.and1.img.Image;
 
 import javax.swing.*;
 import java.awt.*;
 
-public class ImageCellRenderer extends JLabel implements ListCellRenderer
+public class ImageCellRendererSorted extends JLabel implements ListCellRenderer
 {
 
-	public ImageCellRenderer()
+	public ImageCellRendererSorted()
 	{
 		setOpaque(true);
 	}
@@ -17,20 +17,22 @@ public class ImageCellRenderer extends JLabel implements ListCellRenderer
 												  int index, boolean isSelected, boolean cellHasFocus)
 	{
 
+		float similarity = 1.0f;
+
 		if (value instanceof Image)
 		{
-			com.and1.img.Image img = (Image) value;
+			Image img = (Image) value;
 			this.setIcon(new ImageIcon(img.getThumbnail()));
+			similarity = img.getSimilarity();
 		}
 
-		setText(value.toString());
+		setText(value.toString() + " Similarity: " + similarity);
 
 		Color background;
 		Color foreground;
 
 		// check if this cell represents the current DnD drop location
 		JList.DropLocation dropLocation = list.getDropLocation();
-
 		if (dropLocation != null && !dropLocation.isInsert() && dropLocation.getIndex() == index)
 		{
 
@@ -48,8 +50,8 @@ public class ImageCellRenderer extends JLabel implements ListCellRenderer
 		}
 		else
 		{
-			background = Color.WHITE;
-			foreground = Color.BLACK;
+			background = new Color(similarity, similarity, similarity);
+			foreground = new Color(33, 83, 134);
 		}
 
 		setBackground(background);
